refactor: 경매 등록 페이지 로직 분리 및 div 시멘틱 태그로 교체, 누락된 button size 값 추가, errormessage 컴포넌트 분리
