+++ packages/ui/src/utils/getToday.ts
@@ -1,0 +1,4 @@
+export const getToday = (): string => {
+  const today = new Date();
+  return today.toISOString().split('T')[0] as string;
+};
